windows: Child's crash hangs parent.
Parent process must wait also for child process and not only for
interruption signal. Crash of child process won't trigger interrupt
signal and as a result parent will hang. Also in case of crash of child,
parent needs to internally call stop of it's service. Additionally,
sending error from onStop func would not stop whole service. Because of
that now we are logging this error and stopping will always stop the
service.